add controllers logics, admin and public
